+++ R/timeseries_generator.R
@@ -1,73 +1,104 @@
-#' @title Time series data generator for training and testing
+# series_generator
+#-------------------------------------------------------------------------------
+
+#' @title Data generator for training or testing a time series Keras model
 #'
-#' @description A data generator function for training and testing
+#' @description A data generator function for training or testing
 #'  a supervised time series Keras model
 #'
 #' @family time series generators
 #'
-#' @param data A data frame containing the time series data.
-#' @param y Numeric or character vectors that identify
-#'  which column(s) in the \code{data} is/are the target(s).
-#' @param x Numeric or character vectors that identify
+#' @param data A \code{data.frame} or \code{matrix} containing
+#'  the time series data.
+#' @param x \code{numeric} or \code{character} vectors that identify
 #'  which column(s) in the \code{data} is/are the feature(s).
-#' @param lookback A numeric vector which identify
+#' @param y \code{numeric} or \code{character} vectors that identify
+#'  which column(s) in the \code{data} is/are the target(s).
+#' @param lookback A \code{numeric} vector which identify
 #'  the number of lookback period.
-#' @param timesteps A numeric vector which identify
+#' @param timesteps A \code{numeric} vector which identify
 #'  the number of timesteps length.
-#' @param start_index A numeric vector which identify
+#' @param start_index A \code{numeric} vector which identify
 #'  the first row index of the the \code{data}
 #'  that would be included as a sample.
-#' @param end_index A numeric vector which identify
+#' @param end_index A \code{numeric} vector which identify
 #'  the first row index of the the \code{data}
 #'  that would be included as a sample.
-#' @param batch_size A numeric vector which identify
+#' @param batch_size A \code{numeric} vector which identify
 #'  how many observation is considered as a batch.
-#' @param return_target A boolean indicating if the generator.
+#' @param return_target A \code{boolean} indicating if the generator.
 #'  should return the target or not.
+#' @param prep_funs A custom function for preprocess the \code{data}
+#'  on the fly, optional.
 #'
 #' @return If \code{return_target} is set to \code{TRUE},
-#'  the generator will returning and object of \code{list} containing
-#'  a batch of \code{x} and \code{y arrays}.
+#'  the generator will returning an object of \code{list} containing
+#'  a batch of \code{x_array} and \code{y_array}; otherwise,
+#'  it only returns an object of \code{list} containing
+#'  a batch of \code{x_array}
 #'
 #' @export
 
 # time series generator
 series_generator <- function(
 
-  data, y, x, lookback, timesteps,
+  data, x, y, lookback = 1, timesteps = 1,
   start_index = NULL, end_index = NULL,
-  batch_size, return_target = TRUE
+  batch_size = 32, return_target = TRUE,
+  prep_funs = NULL
 
   ) {
-
+  
+  # stop if data is not a proper object
+  if (!inherits(data, c("data.frame", "matrix")))
+    stop("'data' must be an object of 'data.frame' or 'matrix'")
+  
   # check start & end index
   if (is.null(start_index)) start_index <- 1
   if (is.null(end_index)) end_index <- nrow(data)
-
+  
+  # set some global params
+  n_col_x <- length(x)
+  n_col_y <- length(y)
+  
   # start iterator
   i <- start_index
 
+  # return an iterator
   function() {
-
+    
     # reset iterator if already seen all data
     if ((i + batch_size - 1) > end_index) i <<- start_index
 
-    # y rows
+    # iterate current batch's target rows
     y_rows <- c(i:min(i + batch_size - 1, end_index))
-
+    
     # update to next iteration
     i <<- i + batch_size
 
-    # arrays dim
-    x_dim <- c(length(y_rows), timesteps, length(x))
-    if (return_target == FALSE) y_dim <- c(length(y_rows), 1, length(y))
+    # current batch's number of sample
+    n_sample <- length(y_rows)
+    
+    # current batch's row index
+    batch_start <- y_rows[1] - lookback - timesteps + 1
+    batch_end <- y_rows[n_sample]
+    
+    # subset the data into current batch
+    batch <- data[batch_start:batch_end, ]
+    
+    # adjust y rows to current batch's index
+    y_rows <- c((nrow(batch) - n_sample + 1):nrow(batch))
 
-    # array containers
-    x_array <- array(0, dim = x_dim)
-    if (return_target == FALSE) y_array <- array(0, dim = y_dim)
+    # preprocess the batch
+    if (!is.null(prep_funs)) batch <- prep_funs(batch)
+    if (is(batch, "data.frame")) batch <- data.matrix(batch)
 
+    # create container arrays
+    x_array <- array(0, dim = c(n_sample, timesteps, n_col_x))
+    if (return_target) y_array <- array(0, dim = c(n_sample, n_col_y))
+    
     # fill the container
-    for (j in 1:length(y_rows)) {
+    for (j in 1:n_sample) {
 
       # adjust x row by lookback
       x_row <- y_rows[j] - lookback
@@ -79,83 +110,107 @@ series_generator <- function(
       )
 
       # fill the arrays
-      x_array[j, , ] <- data[x_indices, x]
-      if (return_target == FALSE) y[j, , ] <- data[y_rows[j], y]
+      x_array[j, , ] <- batch[x_indices, x]
+      if (return_target) y_array[j, ] <- batch[y_rows[j], y]
 
     }
 
     # return the batch
-    if (return_target == TRUE) list(x)
+    if (!return_target) list(x_array)
     else list(x_array, y_array)
 
   }
 
 }
 
-#' @title Time series data generator for forecasting
+# forecast_generator
+#-------------------------------------------------------------------------------
+
+#' @title Data generator for forecasting using time series Keras model
 #'
-#' @description A data generator function for forecasting
+#' @description A data generator function for forecasting using
 #'  a supervised time series Keras model
 #'
 #' @family time series generators
 #'
 #' @inheritParams series_generator
 #'
-#' @return The generator will returning and object of \code{list} containing
-#'  a batch of \code{x} arrays.
+#' @return The generator will returning an object of \code{list} containing
+#'  a batch of \code{x_array}.
 #'
 #' @export
 
 # time series forecast generator
 forecast_generator <- function(
 
-  data, x, timesteps, start_index = NULL, end_index = NULL, batch_size
+  data, x, timesteps = 1,
+  start_index = NULL, end_index = NULL,
+  batch_size = 32, prep_funs = NULL
 
   ) {
-
+    
+  # stop if data is not a proper object
+  if (!inherits(data, c("data.frame", "matrix")))
+    stop("'data' must be an object of 'data.frame' or 'matrix'")
+  
   # check start & end index
   if (is.null(start_index)) start_index <- 1
   if (is.null(end_index)) end_index <- nrow(data)
 
+  # set some global params
+  n_col_x <- length(x)
+
   # start iterator
   i <- start_index
 
+  # return an iterator
   function() {
 
     # reset iterator if already seen all data
     if ((i + batch_size - 1) > end_index) i <<- start_index
 
-    # sample rows
-    sample_rows <- c(i:min(i + batch_size - 1, end_index))
-
+    # iterate current batch's target rows
+    x_rows <- c(i:min(i + batch_size - 1, end_index))
+    
     # update to next iteration
     i <<- i + batch_size
 
-    # samples dims
-    samples_dim <- c(length(sample_rows), timesteps, length(x))
+    # current batch's number of sample
+    n_sample <- length(x_rows)
+    
+    # current batch's row index
+    batch_start <- x_rows[1] - timesteps + 1
+    batch_end <- x_rows[n_sample]
+    
+    # subset the data into current batch
+    batch <- data[batch_start:batch_end, ]
+    
+    # adjust x rows to current batch's index
+    x_rows <- c((nrow(batch) - n_sample + 1):nrow(batch))
 
-    # array container for sample
-    samples <- array(0, dim = samples_dim)
+    # preprocess the batch
+    if (!is.null(prep_funs)) batch <- prep_funs(batch)
+    if (is(batch, "data.frame")) batch <- data.matrix(batch)
+
+    # # create container array
+    x_array <- array(0, dim = c(n_sample, timesteps, n_col_x))
 
     # fill the container
-    for (j in 1:length(sample_rows)) {
+    for (j in 1:length(x_rows)) {
 
-      # sample row
-      sample_row <- sample_rows[j]
-
-      # select sample indices according to timesteps
-      sample_indices <- seq(
-        from = sample_row - timesteps + 1,
-        to = sample_row
+      # select x indices according to timesteps
+      x_indices <- seq(
+        from = x_rows[j] - timesteps + 1,
+        to = x_rows[j]
       )
-
-      # fill the sample array
-      samples[j, , ] <- data[sample_indices, x]
+      
+      # fill the arrays
+      x_array[j, , ] <- batch[x_indices, x]
 
     }
 
     # return the batch
-    list(samples)
+    list(x_array)
 
   }