+++ tests/testthat.R
@@ -1,0 +1,15 @@
+# prepare environment
+#--------------------
+
+# clear environment
+rm(list = ls())
+
+# load libs
+library(testthat)
+library(kerasgenerator)
+
+# perform test
+#--------------------
+
+# launch all test
+test_check("kerasgenerator")