+++ NEWS.md
@@ -1,3 +1,12 @@
+# kerasgenerator 0.1.1
+
+Time series generator functions now have a proper basic functionality and documentation. Some changes related to time series generator:
+
+* Added a [vignette](/articles/timeseries).
+* Added `prep_funs` argument, which accept a custom function for preprocessing on the fly.
+* Added a error message if not passing a proper data frame or matrix.
+* Added autopreprocess from `data.frame` to `matrix`.
+
 # kerasgenerator 0.0.1
 
 The first version of the package.